+++ src/main/java/com/breckner/happyshop/domain/model/CartItem.java
@@ -3,7 +3,7 @@ package com.breckner.happyshop.domain.model;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Value(staticConstructor = "of")
 @EqualsAndHashCode
@@ -11,5 +11,5 @@ public class CartItem {
 
     CartItemId id;
     Product product;
-    BigInteger quantity;
+    BigDecimal quantity;
 }
